Makes all javascript source "prettier" and enables docs.
The prettier command is used to format javascript source code into a widely agreed upon standard form, so that the developer doesn't need to worry about keeping the code's formatting correc. This commit is the result of running prettier on all the .js files.

This commit also adds esdoc generation of documentation to the package publication process.